refactor(factory): rename BaseFactory to Factory and remove Factory class in favor of factory functions
